Extract haret-cli-client into a lib and bin crate
This starts pulling apart haret (#115), specifically cli client
binary into it's own module. The main reason to do this is for a
few reasons. First, it allows us to start framing out a higher
level library interface for Haret (haret-client). Second, it allows
us to shave off some dependencies if we only need a subset for a
particular application. haret-client will need a lot more attention
over time, since right now it just responds with string output.

Note that I've added a .gitignore to `haret-client`, as the
standard practice in the community is to only lock down
dependencies in the application crates, but leave it up to the
library consumers to decide what dependency versions they want
to use.
