+++ .travis.yml
@@ -4,3 +4,10 @@ rust:
   - stable
   - beta
   - nightly
+script:
+  - cargo build --verbose
+  - cargo test --verbose
+  - (cd haret-client && cargo build --verbose)
+  - (cd haret-client && cargo test --verbose)
+  - (cd haret-cli-client && cargo build --verbose)
+  - (cd haret-cli-client && cargo test --verbose)
