+++ haret-cli-client/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "haret-cli-client"
+version = "0.1.0"
+authors = ["Andrew J. Stone <andrewstone@vmware.com>"]
+description = "A strongly consistent distributed coordination service"
+repository = "https://github.com/vmware/haret"
+keywords = ["distributed", "coordination", "strong consistency"]
+license = "Apache-2.0"
+
+[dependencies]
+haret-client = { path = "../haret-client" }
+uuid = { version = "0.5", features = ["serde", "v4"] }
+lazy_static = "0.1"
+
+[[bin]]
+name = "haret-cli-client"
+path = "src/main.rs"
+test = false
