+++ haret-client/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "haret-client"
+version = "0.1.0"
+authors = ["Andrew J. Stone <andrewstone@vmware.com>"]
+description = "A strongly consistent distributed coordination service"
+repository = "https://github.com/vmware/haret"
+keywords = ["distributed", "coordination", "strong consistency"]
+license = "Apache-2.0"
+
+[dependencies]
+haret = { path = ".." }
+protobuf = "1.0.24"
